requirements.txt
#requirements-core
pandas==2.2.3
numpy==1.26.4
scipy==1.15.2
joblib==1.5.0
#requirements-ml
scikit-learn==1.3.2
statsmodels>=0.14.0
PyPortfolioOpt==1.5.0
riskfolio-lib==7.0.1
cvxpy==1.6.5
ecos==2.0.14
osqp==1.0.3
gymnasium==1.1.1
#requirements-data
requests==2.32.3
beautifulsoup4==4.13.4
feedparser==6.0.11
pandas-datareader>=0.10.0
#requirements-dev
jupyterlab==4.4.2
ipykernel>=6.0.0
tqdm==4.67.1
python-json-logger>=2.0.0
APScheduler==3.11.0
python-dotenv>=1.0.0
pydantic==2.7.4
sqlalchemy==2.0.6
#requirements-quant
yfinance>=0.2.18
FinMind==1.7.8
alpha_vantage==3.0.0
backtrader==1.9.78.123
freqtrade==2025.4
freqtrade-client==2025.4
tensortrade==1.0.3
pyalgotrade==0.20
qlib==0.0.2.dev20
gym-trading==0.0.2
finrl==0.3.7
autotrader==0.12.1
ccxt==4.4.80
optimus==2.1.0
jesse==0.33.1
ta-lib==0.6.3
#requirements-visual
matplotlib==3.10.1
seaborn==0.12.2
plotly>=5.0.0
ipywidgets==8.1.7

